Ajax loading places in Customer Create and Edit
