+++ Termoservis/Termoservis.Web/Models/Customer/CustomerFormViewModel.cs
@@ -38,14 +38,6 @@ namespace Termoservis.Web.Models.Customer
         public string ActionName { get; set; }
 
 		/// <summary>
-		/// Gets or sets the available places.
-		/// </summary>
-		/// <value>
-		/// The available places.
-		/// </value>
-		public IEnumerable<Place> AvailablePlaces { get; set; } = new List<Place>();
-
-		/// <summary>
 		/// Gets or sets the available countries.
 		/// </summary>
 		/// <value>
@@ -71,6 +63,15 @@ namespace Termoservis.Web.Models.Customer
 		/// </value>
 		[DisplayName("Mjesto")]
 		public int? CustomerPlaceId { get; set; }
+
+        /// <summary>
+        /// Gets or sets the name of the customer place.
+        /// </summary>
+        /// <value>
+        /// The name of the customer place.
+        /// </value>
+        [DisplayName("Mjesto")]
+        public string CustomerPlaceName { get; set; }
 
 		/// <summary>
 		/// Gets or sets the country identifier.
