+++ Termoservis/Termoservis.Web/Termoservis.Web.csproj
@@ -103,8 +103,8 @@
       <HintPath>..\packages\Unity.AspNet.WebApi.4.0.1\lib\net45\Microsoft.Practices.Unity.WebApi.dll</HintPath>
     </Reference>
     <Reference Include="Mindscape.Raygun4Net, Version=5.3.1.0, Culture=neutral, processorArchitecture=MSIL">
+      <SpecificVersion>False</SpecificVersion>
       <HintPath>..\packages\Mindscape.Raygun4Net.Core.5.3.1\lib\net40\Mindscape.Raygun4Net.dll</HintPath>
-      <Private>True</Private>
     </Reference>
     <Reference Include="Mindscape.Raygun4Net.Mvc, Version=5.3.1.0, Culture=neutral, processorArchitecture=MSIL">
       <HintPath>..\packages\Mindscape.Raygun4Net.Mvc.5.3.1\lib\net40\Mindscape.Raygun4Net.Mvc.dll</HintPath>
@@ -259,7 +259,12 @@
     <Compile Include="App_Start\WebApiConfig.cs" />
     <Compile Include="Controllers\AccountController.cs" />
     <Compile Include="Controllers\CustomersController.cs" />
+    <Compile Include="Controllers\v1\Select2\ISelect2Item.cs" />
+    <Compile Include="Controllers\v1\Select2\ISelect2RequestData.cs" />
     <Compile Include="Controllers\v1\PlacesApiController.cs" />
+    <Compile Include="Controllers\v1\Select2\Select2Item.cs" />
+    <Compile Include="Controllers\v1\Select2\Select2RequestData.cs" />
+    <Compile Include="Controllers\v1\Select2\Select2Response.cs" />
     <Compile Include="Models\Customer\CustomersSearchResult.cs" />
     <Compile Include="Controllers\HomeController.cs" />
     <Compile Include="Controllers\ManageController.cs" />
