+++ Termoservis/Termoservis.Web/Views/Customers/_CustomerFormPartial.cshtml
@@ -41,7 +41,7 @@
             <div class="form-group row">
                 @Html.LabelFor(model => model.CustomerPlaceId, htmlAttributes: new {@class = "form-control-label col-md-2" })
                 <div class="col-md-10">
-                    @Html.DropDownList(nameof(Model.CustomerPlaceId), new SelectList(Model.AvailablePlaces, "Id", "Name"), htmlAttributes: new {@class = "form-control select2-init", tabindex = "3"})
+                    @Html.DropDownList(nameof(Model.CustomerPlaceId), new SelectList(Model.CustomerPlaceId.HasValue ?  new List<SelectListItem> {new SelectListItem{Text = Model.CustomerPlaceName,Value = Model.CustomerPlaceId.Value.ToString(),Selected = true}} : new List<SelectListItem>(),"Value", "Text"), htmlAttributes: new {@class = "form-control select2-init", tabindex = "3"})
                 </div>
             </div>
